[debug] Add helper tool crashpad_startup
This is intended to be run from devmgr, eventually at the same point as
crashlogger is run today. At the moment, it's run later once /system and
/data are available.

This tool is mostly a configuration shim to configure crashpad_handler
(the tool from the Crashpad repo) in a Fuchsia-specific manner
(target directories, etc.)

For the devmgr hookup, see: https://fuchsia-review.googlesource.com/#/c/zircon/+/151636.

DX-21 #comment

Change-Id: I621121d11969ce1078b9ff02a5af3a26966c9937