[sketchy] Apply stroke fitter on buffered input points.
The stoke fitter didn't work before, because fitter::Fit() would be
called on every Stroke::Extend(). And when fitting two points, the
fitter will simply draw a line.

In the new design, the fitter no longer owns a path. Instead, it will
own the points that are not plenty enough and should be fit again when
the new points come. When there're plenty enough points, however, it
will pop as many points as it can, to avoid fitting too many points
when the curve gets longer.

Change-Id: I9cd5b64a6471db4f792f02c2c672e6378bc906e6