[audio][mixer] Mixer optimizations
With template specializations, the case of "modulo==0" is
optimized so that recently-added modulo arithmetic is
avoided. For those cases, with the admittedly less-
sophisticated performance-profiling method that I've been
using, this is a perf win of 5-35% -- for those cases,
essentially returning performance to what it was before
the position-precision fix went in.

Change-Id: I190aaa2d10aae856bc29e58cd2653fa191622619