[threads] Make zxr_thread_t opaque.
Previously the definition for zxr_thread_t was in a public header, in
spite of accessor functions that implied it should be internal. The
definition of this struct also depended on C atmoics which prevented
this file's inclusion from any C++ source file.

Making this struct opaque allows the removal of the atomics dependency,
better encapsulation, and inclusion in C++ sources.

This could have been donw with an allocate/deallocate function but the
current uses all allocate the structure by value and this seems to be
appropriate for these uses. Instead, a structure of the correct size is
created for public consumption, enforced by a static assert.

Change-Id: I2c72b8ed990f5e09af2cef93846062fc7eec932f