[benchmarks] Fix test names as shown in Catapult
If the name of a results file ends with ".X", the fuchsia_perf recipe
strips that suffix off.  That means that "zircon.perf_test" appears just
as "zircon".  Work around this by adding a ".json" suffix to all the
results filenames so that that part gets removed and the
".perf_test" part doesn't get removed.

Change-Id: I1872dfddd67181046fb8e16f2baa3ac59eec1a85