[tests] Test components are more uniform.
All test components defined in integration test suites
are uniformly implemented by a main class called TestApp.

The TestApp is defined in the anonymous namespace, not in the
modular namespace. This makes it more apparent that we take
an outside look on the modular API here.

TestPoint is using-declared at the top of the file.

TestPoints are defined before the method that passes them, even
if the method is public.

All TestApp classes are defined moveonly.

Change-Id: Ica81cc2a32dcab5c9cb5c900629acb85386f0f32