[guestmgr] Add ManagedSocketEndpoint.
This change converts the |VsockEndpoint| class into an abstract
class with 2 concrete variants:

  * RemoteVsockEndpoint - Maintains the previous behavior; sends all
        requests to a single |SocketAcceptor| and requires the consumer
        to manage out-bound ports.

  * HostVsockEndpoint - Maintains a map of per-port listeners that can
        be registered using the |ManagedSocketEndpoint.Listen| IPC
        method. Additionally this component will manage the set of
        in-use ports and allocates ephemeral ports for callers. As such
        |ManagedSocketEndpoint.Connect| method only requires specifiying
        the destination (cid,port) tuple.

GuestEnvironment is updated to vend out a |ManagedSocketEndpoint| so
that Fuchsia programs can more cooperatively share the host socket
endpoint.

The `guest socat` command has been updated to remove the need to
provide the 'host port' for out-bound connections as this is now
managed directly by the guestmgr.

Change-Id: I42e1d8d0e9175ee824ce439fbde4cd974e8f42bf