[wlan][wlanstack] BSS compatibility check
Fuchsia's security protocol support evolves over time.
In the meantime, this CL provides an indication on each BSS
if it is compatible to Fuchsia's operation or not.

AP FIDL struct is extended to support the compatibility marker.
"wlan" CLI tool shows compatibilty info along with the scan results.

Test:
NET-481 #done

Change-Id: I51e46a805ca0e41f31327ff170fa45249f553c4a