[wlanstack] Give up unless auth_failure_timeout
When SME gets an authentication rejection from the AP, it will continue
to try to connect to the rejecting AP. This creates an infinite loop and
wlanstack cannot take new commands because retrying with existing BSSID
has higher priorit.

Because SME does not support cancelling current command, the only way to
allow user to intervene is to clear the existing Config so that it can
take other commands in ScanState.

Meanwhile, if and only if authentication fails due to timeout, continue
to authenticate as before.

Test: Verified that SME stopped trying after failing to join an AP with
MAC filter rejecting the client WLAN MAC address.

NET-844

Change-Id: Iae8c7aae635038c6809861465a3f3afae3fff602