[kernel][syscalls] sanitize ZX_PROP_NAME
Modify the setter functions for ZX_PROP_NAME to disregard data from
userspace which occurs after the trailing NUL character. Additional data
beyond NUL is garbage from userspace and shouldn't be stored in the
kernel.

Modify the getter functions to completely sanitize passed in memory
before returning it to userspace or other parts of the kernel.

Add additional tests to exercise this code and prevent regressions.
Fixup unittests (added in commit 22efc1d716) which asserts that the
buffers aren't cleared.

Fixes ZX-2048, which appears to have been introduced in commit
22efc1d716.

Change-Id: Ifd29767bb4ab7b1866ad90503fab8fd1f36ca543