[vim2][mailbox][scpi] Introduce Mailbox & SCPI Protocol
Breaking down the fan-control thermal prototype driver
into 3 parts.

1st layer would be the Mailbox driver which would
provide communication between the two on-board
processors.

2nd layer would be the SCPI driver (System Control and
Power Interface) which would use the mailbox protocol
to set/get system configuration such as frequency/voltage
and thermal parameters.

3rd layer would be the Fan-control driver which uses
the SCPI protocol to get the temperature and controls
the fan via the GPIO protocol based on it.

(TODO): Expose the fan-ctl & SCPI interface as IOCTL's in a later CL

Change-Id: Ia0175fdfe76048d7cb789e3898e50d9cffb3cee3