[kernel][channel] Reduce unnecessary context switches using AutoReschedDisable
On a single-CPU system (booting with kernel.smp.maxcpus=1), this
reduces the number of context switches per run from 4 to 2 for the
following performance tests in zircon_benchmarks:

  RoundTrip_BasicChannel_*
  RoundTrip_ChannelPort_*
  RoundTrip_ChannelCall_*
  RoundTrip_Fidl_*

The context switch counts for the other tests are unaffected.

This change will allow some cleanup in follow-on changes: plumbing
such as kWokeThreads can be removed, because it's replaced by using
AutoReschedDisable.

ZX-1472

Change-Id: I5206a71063f7bd3aacb97399748be0e8e68ec151