[debug][devmgr] Create crashsvc as system-level exception handler
crashsvc listens on the root job's exception port and dispatches to
back to devmgr to have it deal with the exception. devmgr in turn
launches an analyzer process to do something with the process that
sustained the exception. This is to allow for customization of the
exception handler, rather than always having crashlogger handle
exceptions. Short doc on rationale at go/crashsvc.

For the time being however, a refactored crashlogger (named
crashanalyzer) is still always used as the analyzer. Followups will
allow for customization of the analyzer.

crashanalyzer is a stripped down copy of crashlogger. For the time being
an #ifdef ENABLE_CRASHLOGGER is available in devmgr for easy switching
on/off, to restore the previous behaviour of registering crashlogger as
the exception handler on the root job.

Change-Id: Id10c52c2a31a5ccfeed9b63eb794559c77592df8