[debugger] Improved breakpoint backend.
Previously, the design required the client to compute the individual
locations for each breakpoint. This simplified the agent but made it
impossible to have more than one breakpoint at an address, and to share
state like hit counts across locations.

This new design provides all processes and addresses that a breakpoint
applies to from the client. A new Breakpoint object manages this state.
The DebuggedProcess objects are in charge of mapping these to
ProcessBreakpoint objects. ProcessBreakpoints are mostly the same as
before but they can now reference multiple Breakpoints.

Fixes a shutdown crash when a breakpoint is present because the target
destruction notification was not fired.

Fixed an error processing LLVM symbols.

Change-Id: I99a7454e229b6448a5977f778095a9511ad35ead