[garnet][zxdb] address space
This CL introduces a new debugger command: aspace
which returns the address space of the process
being debugged. You can use 'aspace' or 'as'

The result with no params is the dump of the
vmars in the process, starting from the topmost
and traversed in-order. Internally the is
a |depth| value which goes from 0 to N as the
vmars can be subdivided.

There is no attempt to keep this information
on the debug_agent side or on the debug side, further
improvements would include having this info
automatically synched at certain points since
it provides info that cannot be obtained otherwise

Change-Id: I3c18bd0bf9285b9efaccd5eba1dca5be33f6f627