[audio] Add a device token to AudioDevices
Add a 64 bit device token to AudioDevices in the audio_service.  This
token will serve as the unique-per-boot identifier of audio devices
(inputs and outputs) when they are enumerated.  For now, we use the
KOID of the stream channel to the driver as this ID.

Also remove the clang-format-(on|off) directives surrounding the
various lambdas in the code.

Test: Manual testing on NUC with simple_sine and vol.  Checked
      built-in output and USB output.
Change-Id: If28e40baec5a94384590994f05e262701d51ae0a