[media_player] fail early when we exhaust audio renderer VMO space
We have a fixed amount of memory available for payloads of packets
sent to the audio renderer. We're allocating a full second's worth,
which should be enough, so there's a bug to fix there.

The stack dump when this occurs is misleading, because we're trying
to write to a null pointer. This CL checks that pointer after the
allocation call and bails immediately if the allocation fails.

We may need a way to deal with allocation failures gracefully, or we
may just prevent them from happening altogether. In the latter case,
blowing up the process as we do here might be reasonable.

To be clear, this CL doesn't fix anything. It just produces more
comprensible stack dumps.

MTWN-130

Change-Id: Ie2957fdb0acada806b7b3f79806ae48d1193a03c