[story_shell] Route Presentation request
Give StoryShell the ability to request a Presentation service
instance from the UserShell, in the same way as UserShell can
already request a Presentation service instance from DeviceShell.

The ability is provided by way of a method GetPresentation()
on the StoryContext passed to StoryShell, analog to the method
on the UserShellContext passed to UserShell.

The request passed in StoryContext.GetPresentation() is passed
to UserShellPresentationProvider.GetPresentation() implemented
by the UserShell component, as a separate public service.

Added a test case to the story_shell integration test that
exercises the functionality.

Change-Id: Ife520dbb8edd9be9df4b26b60ae40853758e5433