[tools] Always output text no matter what
Currently if a parsing error occurs the symbolizer may give up on a line.
This is suboptimal and the symbolizer should always make a best faith
effort to a) symbolzie the output and b) get as much of the text to the
user as possible. This change closes the last way in which text might
not be output.

Additionally a common occurence is for a line to be partially overlapped
by a kernel line. For instance it might look like "foobar [131.00] ...".
This change treats this as "foobar\n[131.00] ..." instead. This allows
the proper part of the log line to be parsed correctly while allowing
the additional text to not simply be dropped but instead be sent to the
user.

Change-Id: Ia4302ac3b75f7d40684d4afc0079904a9fbdf2bc