[cobalt] Start encrypting observations and envelopes.
A key part of the design of Cobalt is that observations sent to the
Shuffler are doubly-encrypted. The observations themselves are
encrypted to the public-key of the Analyzer and the Envelope
containing many observations is encrypted to the public key of
the Shuffler.

This functionality is implemented and when we run clients on Linux
we use it. But up until now we have not been using it from
Fuchsia. This CL rectifies that.

We now package the public keys of the Shuffler and Analyzer into
the Cobalt process's package and read them at run time.

We also make a second tangential changes in this cl:
We roll Cobalt in order to pick up a build fix that we need:
https://fuchsia-review.googlesource.com/#/c/cobalt/+/152515/


Change-Id: Iff37e9dfb4c57ca800acabb6d3a35394b281792b