[overnet] Initial Endpoint API
Apologies for this being a little bit of a 'paint the owl' change.

Implements an initial endpoint API - which I expect to be the API upon
which the FIDL bindings are programmed.

Adds:
- Optional, a polyfill for std::optional
- Source, analogous to Sink
- ForkFrame, a frame that describes a forking of stream 0
- stream 0 - a control stream between two nodes (the connection stream)
- Linearizer - for reassembling reordered bytes into their correct
  order
- Slice - for carrying ref counted arrays of bytes

Moves some code around:
- SeqNum, NodeId, StreamId move into their own headers and impl files
- Slice moves into its own set of files

I intend to do some further refactoring to RouterEndpoint, and do not
consider it a final API. Particularly, I'd like to raise all memory
allocations along the data path so that they are performed by clients of
this API instead of by the implementation.

Additionally this change does not deal with backpressure or message
acknowledgement/resending - both major and intertwined subjects that
will be the subject of a future CL.

Change-Id: I84ff7a04d387e2652a18f7652ae8a5529a95bae8