[media][tests] Profile integer mixing
This CL profiles the Mix() function for int16 and uint8
input formats, augmenting the existing profiling for float
inputs.

It also adds a display of the first, best, and worst timing,
in addition to the average timing.

Having discovered that after the first few runs the variance
in timing is minimal, this CL also reduces the number of
runs for a configuration from 1000 to 100.

Change-Id: I57699c1d5bd40e0dbe01ebb14c43a69325ca1c62