[fbl] AutoLock: Improve code generated for mutex release
Before this change, the "if (mutex_)" conditional is often not
optimised away by the compiler because the compiler doesn't know that
the mutex pointer passed to AutoLock is non-null.

We can fix this by using a separate bool to track whether the mutex
needs to be released.  This is easier for the compiler to optimise
away.  I checked the generated code:

  Total number of "if" lines from release() (from "objdump -dl"):
    gcc: before 165, after 0, difference -165
    clang: before 64, after 1, difference -63
     * note: clang fails to inline ~AutoLock() (before and after) in 2 cases.

  Total number of conditional jumps (from "objdump -d"):
    gcc: before 20740, after 20643, difference -97
    clang: before 19176, after 19114, difference -62

  Text segment size (from "size" tool):
    gcc: before 1112107, after 1111503, difference -604
    clang: before 1220568, after 1220280, difference -288

(An alternative would be to have different variants of AutoLock with
and without an early-release method.)

Change-Id: I2d74d53e969e4bc5e3e201686a19ed28c974bbe8