[story-state] Drop DONE State.
The DONE ModuleState is used only for signalling in tests, but
not in actual production modules. Hence, drop it.

In tests, it's replaced by a TestStore key kTestShutdown which
DevDeviceShell listens for, and when it observes it calls Shutdown()
the same way it used to when it observed the test story to be DONE.

Collateral cleanup:

* Use calls to modular::testing::Await() throughout tests.

* Unify test timeout in DevDeviceShell, remove repeated code from
  test components.

* Delete CommonDoneModule, as it does the same as CommonNullModule
  now.

MI4-304 #comment Drop DONE State.

Change-Id: Iccd86f25bd76b4443485cf82e287c5f4ab67a57b