[kernel] Don't relocate kernel on secondary CPUs
This change fixes a bug that could cause corruption of global data
structures.

_start is the entrypoint for all CPUs. Prior to this change every CPU
would relocate and patch the kernel. Depending on the link layout the
fixup code may or may not overlap with the BSS.

If they do overlap, then the secondary CPUs may run off and begin
executing part of the BSS as if it were code. In one case this
resulted in secondary CPUs "fault looping".

If they don't overlap, then the result can be more subtle. Even though
the relocation offset is currently 0 (the kernel stays put), this can
lead to bugs because the fixup code is non-atomic.  Imagine that one
CPU is writing to the data segment while a secondary CPU is relocating
the kernel.  For example, CPU 0 is about to update a global variable
located at address A.  Meanwhile, CPU 2, running the relocation code,
loads the value at A, adds 0, but before it can store it, CPU 0's
store completes.  Now when CPU 2 completes its store it stomps what
CPU 0 wrote.

ZX-2140 #done

Change-Id: I90a58ca85866ca47cd2b82f131e0d5e661141793