[auth] Initial Rust implementation of Token Cache
This is intended to be a port of the existing C++ implementation in
bin/auth/cache without introducing significant refactoring. Note this
code is not yet used by the remainder of the system. This is the first
Rust code for the Auth team and for myself so it may need some work.
Once this is in, future CLs will add Rust implementations and tests
for the remaining components of the TokenManager and then replace the
default C++ implementation.

AUTH-45

Change-Id: I8d831105f8152a86374935c5b3d1de7bd2e41902