[media_player] add AsyncNode node model
This CL adds a new node model for the in-proc streaming framework
that will replace all other asynchronous models. At the moment, we're
only using asynchronous models, so it may soon be the only model.

We will probably add a model for all nodes that operate synchronously.

All the asynchronous models were starting to look pretty similar. The
only significant distinctions between them concerned the number of
inputs and outputs. The AsyncNode model handles any number of inputs
and outputs, configured during node setup. The base class is defined
in a way that makes it easy to ignore inputs when writing a source
node, to ignore outputs when writing a sink node and to ignore
input/output indices when there is only one input/output.

The Demand type no longer appears in the definition. At some point,
Demand will be removed entirely from the framework.

The new model also eliminates a problematic pattern that appeared in
the old definitions. Demand was communicated in two ways:

1) When supplying a packet, the method that received the packet would
   use its return value to indicate whether another packet was
   demanded.
2) Demand could be expressed asynchronously with a separate call.

This pattern was used for both inputs and outputs and tended to
create race conditions that were hard to solve. In the new model,
only the separate call is supported, but that call can be made from
the method that receives a packet.

Change-Id: I83b75ee2d6595870ddce37f1ce49000a65118470