[media_player] FidlAudioRenderer fixes
This CL addresses a threading issue in FidlAudioRenderer and adds
some code to detect and document MTWN-130 before it manifests in
a fault.

FidlAudioRenderer is an allocator for upstream nodes (the audio
decoder, specifically). Allocators are required to be free-threaded,
so a mutex is required. Currently, the audio decoder and the audio
renderer use only one thread (for both), so this issue wasn't causing
any failures. The decoders will soon use their own threads, and this
will become important.

The code for detecting MTWN-130 just logs a short report when it
notices that the renderer is holding on to too much data.

Change-Id: I0cac385f9b3dea0f0ec8f02442970bbc1365929e