[x86][page_tables] Remove a use of TA_NO_THREAD_SAFETY_ANALYSIS
Remove a comment which doesn't seem to be accurate: pages() is only
used by pt_pages(), which is only used for tests (and not for
calculating memory usage).

It's not true that it's OK to access a non-atomic field concurrently
without holding a lock: Doing that can run into undefined behaviour,
because a compiler might duplicate reads or write temporary junk
values.

We could use atomic accesses to address that, but since this field is
only used by tests, it's easier to just claim the lock.  This
addresses a potential future correctness gotcha and allows removing an
instance of TA_NO_THREAD_SAFETY_ANALYSIS.

Change-Id: Ieffae384c3727be472fde14911aba356214b55b7