[minfs] [blobfs] Update behavior when FVM slice counts differ.
When FVM reports a lower slice count than Blobfs or Minfs,
automatically fail, since it is possible that one or more of the
slices expected by the file system has data allocated to it.

When FVM reports a higher slice count, always attempt to shrink
the FVM - assume that the file system is the source of truth.
Since those slices aren't expected to be there, we aren't losing
any crucial data by shrinking them.

Change-Id: I44e7cde936b8df435564c86f432a474141ac0e6b