[netstack] Get nodename from device_settings_manager.
Update the ioctl uname invokes: check the device settings service
for a set device name and fall back to a default (here
"fuchsia-unset-device-name") if none is set.

Remove unused GetNodeName FIDL API.
Configure sysmgr to automatically start netcfg, which sets the default
device name to "fuchsia."

I considered returning zx_err_should_wait if the device name was unset,
but it's not clear if it's reasonable to expect the caller to wait long
enough for a network interface to come up, which is the timescale
we'd currently expect before device name is set by netcfg.

In the future the ioctl will go away entirely and uname will call
through to the device settings service from Zircon.

Change-Id: Icb579dc9da42009b75363e6238b81fed57247ba4