[x86][registers] Add API to expose xsave info.
The format of the xsave legacy area is moved to the header and all
values are filled out (previously this was just in registers.cpp and
only mxcsr was defined).

The existing #defines for states were renamed *_BIT_* since they're bit
masks, and a new set of defines for the state indices was added. These
are used for referring to the state number in the new API.

Logic for rounding the xsave area was moved to
x86_extended_register_init_state from thread.cpp. This isn't strictly
necessary (originally this was to support a version of this patch that
was removed) but this seems like a cleaner split of responsibilities. In
support of this, a new define for the required buffer size is added.

Change-Id: I91c4ee2d27419f3b6bb6f23582ea7275a3566396