[machina][vsock] Fix fwd_cnt in credit updates.
This is a free-running counter that is used by the guest to calculate
available space in the buffer:

  /* tx_cnt is the sender’s free-running bytes transmitted counter */
  u32 peer_free = peer_buf_alloc - (tx_cnt - peer_fwd_cnt);

This exposes another issue where guest transmits may stall waiting on a
credit update. We'll need to add support for sending unsolicited credit
updates back to the guest when socket space becomes available but this
at least prevents the guest from sending buffers that would overflow
the socket.

Change-Id: Ic32df5ccc05461f618a27fe20d53cf5f7c1f1ef1