[netstack] Add DHCP Server
This change adds a DHCP Server type and begins to implement its handling
of DHCP Discover messages. Currently, the implementation can allocate an
IP address and issue it in a DHCP Offer message, along with the the
required set of DHCP options. The complete address selection logic will
be iterated upon in future changes, as will option handling.

test: passes unit tests

NET-752 # progress

Change-Id: I5e7c5310f27198ba735d954ee2a9459536a5dbc2