[perftest] Add support for multi-step tests
See perftest.h for docs.  An example of a multi-step test is running
malloc()+free() repeatedly, and recording the times taken by malloc()
and free() (added in malloc-test.cpp).

This adds some complexity to the logic in runner.cpp for recording the
timestamps.

For the time being, this doesn't involve any change to the schema of
the JSON results file.  For a multi-step test, we report the results
of each step as if it were a separate test case.  e.g. The
"MallocFree" test produces results for "MallocFree.malloc" (times for
the malloc() calls) and "MallocFree.free" (times for the free()
calls).

ZX-2070

Change-Id: I4d90cbee51920d039916632e3ea3153bb686a805