[wlan][wlanstack] Spawn additional goroutine
If CLI tries to query the status of wlanstack while it is scanning,
CLI blocks until the scan finishes.

It turns out with the current golang FIDL bindings implementation,
goroutines have to be manually created. So the solution is to create
another goroutine so that wlanstack can handle 2 commands at the same
time.

Test: wlan status returns immediately when wlanstack is scanning, on
both NUC and Eve.

NET-727 # done

Change-Id: I098b8af98b0527b3ddea790678308685bc9d0c40