[bt][hci] Simultaneous SequentialCommandRunner
SequentialCommandRunner is extended to allow for gated groups of
commands.  If commands do not depend on the previous commands having
succeeded, the wait parameter of QueueCommand can be set to false.

If any command fails, the entire sequence fails and whether the
remaining command callbacks are run is unknown.

Any command which has been sent to the controller will result in it's
callback being called with a Command Complete (or failing Command Status).

NET-729 #comment
Test: bt-host-unittests

Change-Id: I998e40d466aba1da963cd8bd0943f90a5d455869