[audio][mixer] Cleanup throttle upon format change
When clients change the format for an audio renderer, existing checks
ensure that there are no remaining enqueued media buffers. We then
unlink all connected outputs, before constructing new links in the
new format. Recent changes with regard to the throttle output begin
to treat that output like any other destination. However, there is
still a special member within AudiorendererImpl, and this needs to
be explicitly cleared when the more general Unlink() is called.

MTWN-132 #done
Test: Subsequent CL (moving existing tests and adding new binary)

Change-Id: Ice457e2bcf1c9e66c9bd2ba0a892e1a182fa69c3