[amber] Move amber daemon errors along to amber_ctl
Currently amber daemon errors result in a closed zx channel which
causes amber_ctl to report a channel error. The user also never
sees the daemon error. With this cl, the daemon uses a signal to
inform amber_ctl that an error has occurred, and sends the error
string along through the channel for amber_ctl to print out.

Change-Id: I530abe972f176f35d1e534a3e581b02081a877bf