[zbi] Replace unused word in zbi_kernel_t with reserve_memory_size field
This informs the boot loader of the kernel's requirements for
scratch memory space available immediately after its load image.
This is where the kernel's .bss (and the early copy of the KASLR
fixup code on x86) goes.  The boot loader is responsible for not
placing anything else (like the ZBI items or other reserved
regions) too close to the end of the kernel's load image.

Change-Id: I34328c2e1b7063ebe6d56558ba6774f92db6887f