[media_player] fix sink connect problem reporting
This CL fixes the way connection problems are reported from a sink
segment to Player.

SinkSegment::Connect has a completion callback, but it wasn't
reporting the outcome of the connection attempt. Connection can fail
when the sink segment is unable to find a decoder for the stream type
it is given.

In addition, Player was reporting connection failures is a peculiar
way, when the sink segments have the ability to report problems
themselves.

This CL adds a Result to the Connect callback so the sink segment can
let Player know whether the operation succeeded or failed.

It also moves the responsibility for connection error reporting to
the sink segment, simplifying Player's logic.

Change-Id: I28f5f302e08cdc4fc386018ca038be6a08dbbc7e