[core][devmgr] Set ZX_PROP_NAME on driver module VMOs
With dlopen_vmo of a module with no DT_SONAME (such as a driver), the
dynamic linker has no name to report in logging messages except what
it can derive from having just the VMO handle.  So it synthesizes a
name to use in messages, and embeds the VMO's ZX_PROP_NAME string in
this name if it can.  Right now drivers tend to be VMO files from the
BOOTFS and those all get the name "bootfs", which is not very useful.
So explicitly set the VMO name to the driver module file's basename.

Change-Id: I4a7dbd0730a05471ff09c921118b85fdad207396