[media][tests] AudioServer API tests
This CL introduces a new audio_server_tests binary focused on
API-level testing of audio_server, starting with AudioRenderer2.

This test binary was created specifically to verify the fix for
MTWN-132, in AudioRenderer2 (and AudioRenderer). For this reason,
only four AudioRenderer tests have been added -- those directly
related to the fix, targeting AudioRenderer2 and AudioRenderer2Sync.
As a prerequisite, tests for AudioServer and AudioServerSync have
also been added, validating those interfaces in full.

Change-Id: I7b808889dc7794eba0f06f7cdac3207950186457