[modular] Don't assert existence of chain
When connecting an intent parameter that is a connection between two
links, the framework was verifying the existence of the source link
prior to creating the link. This isn't necessary, as a new link can
be created.

There is a remaining issue related to making sure that it's clear
how to construct an intent "correctly", but that will be solved
separately.

This also uses updates a clone to use fidl::Clone() function instead
of the Clone method, where a module manifest could be null.

MI4-992 #done

Change-Id: I7aee3b245401bff45c1e184fbda0e1c13c51f1a8