[tools][symbolize] Make text parser more robust
Currently text will never parse past "{{{" or "\033[".
This causes issues if you have partial versions of either
these embeded in otherwise meaningful lines. So we'd like
to not give up on the whole line just because one part
looks a bit funny.

This change makes text consume everything upto a regular
expression which matches the parts of text that can actully
be parsed.

Change-Id: I5ed22d6d8cf5cc53fe5dcc5687ec87e13e9144e9