[agent_runner] Initialize delegate_ as null.
While using the Scudo allocator, I noticed that it's possible for the
DCHECK in delegate_ to fail. This occurs because Initialize assumes
delegate_ will be zero.

Thus, we initialize delegate_ to nullptr in the constructor.

(An alternative fix for this would be to simply remove the DCHECK, since
there's a commented warning in agent_runner_storage_impl that delegate_
is only valid *after* Initialize() is called, but I suspect this
solution is less failure-prone.)

Change-Id: I5cc8b20a351d0ae1b16f058ecd4abad08858bda4