[utest][core][memory-mapping] Use less wildly undefined behavior
The PROT_EXEC test was based on using a lot of deeply undefined
behavior in C++ code that is fundamentally incompatible with
anything even slightly trying to enforce the language semantics.
Rewrite it using a tiny bit of assembly code that keeps what the
language doesn't understand entirely out of the language.

Change-Id: I9a9076086dfa8713fc21f49186beb3b9b6e9617e