[debugger] Initial symbol indexing implementation.
Adds a class that provides a simple symbol index for a module (shared
library or executable). This code is not yet hooked up to the debugger.

A test program is added that is compiled for the target and has specific
symbols. The index loading and lookup is tested by loading the resulting
binary.

Currently only very simple exact-match symbol lookup is possible, and
template parameters are not handled. This can be expanded in the future.

Change-Id: Iaf106481d3f685785c542e92e11f8caa94c2519f