[media_player] delete unneeded files
We no longer need the MediaProducer sink or the MediaConsumer source
or the stages that host them. FIDL formatting is also no needed.

The MediaProducer/Consumer sink/source may be needed in some form in
the future for moving elementary streams between processes. In that
case, they need to be resurrected and modified to use the AsyncNode
model. For the moment, we have no easy way to test a conversion to
AsyncNode, so the producer/consumer should go away.

Change-Id: Ib3cf49a5bfa5e8b8a46714461ba54a82d62c1ea7