[media_player] remove postive/neutral demand distinction
The in-proc framework distinguished positive demand, in which an
input needs a packet soon in order for the graph to function
properly, from neutral demand, in which the input can accept a
packet but can wait until parallel demand causes a packet to be
produced. This makes a certain amount of sense when one leg of
a graph is driving demand (an audio renderer, e.g.) and another
leg isn't (a file writer, e.g.). If the file writer were to
positively demand packets as fast as it could write them, that
would drive the entire graph to end-of-stream faster than real
time.

At the moment, we aren't building any graphs that have this
passive/active leg asymmetry. When we do, the positive/neutral
distinction can be established statically when the graph is built.
We never dynanmically transition between positive and neutral
demand.

Change-Id: Ib3d4c61c212babfb6c1b5e9f0670231295f55e09