[media_player] simplify prepare/unprepare and remove engine.*
This CL simplifies then way prepare and unprepare are done. Both
operations proceed upstream in a breadth-first walk of the graph.
Previously, the individual nodes got to decide which of their
inputs would be traversed. We always traverse all the inputs, so
that feature is no longer needed.

The code for these graph walks used to be implemented in the Engine
class. That class used to have many responsibilities, including
deciding which stages/nodes would be updated in what order. Prior
to this CL, Engine had shed all of its state and become a place to
put code that really belongs in the Graph. This CL moves all that
code to Graph and deletes engine.*.

Change-Id: Icf6890700ba1794b73b632e6306424969dcb6c86