[modular][puppetmaster] Provide PuppetMaster from user_runner.
The StoryCommandExecutor in use would currently crash on any request,
since it does not have any command runners. However, this allows us to
start using new commands immediately as they are implemented.

This is a progress CL.

Also:
* Add a simple function to generate the production version of the
StoryCommandExecutor. This will be the entrypoint for any FIDL services
owned by user_runner that want to make changes to a story (ie,
ModuleContextImpl).

TESTED=Not yet.

Change-Id: Ib7526d34ee3b4c338d1394c9cf888850caf076d9