[wlan][mlme] Fix early PS-POLL frames
If the Beacon has the group addressed traffic bit set in its TIM,
and the client was not yet associated, it used AID 0 to check
for available traffic and as a result attempted to poll frames
whenever there were multicast ones buffered.

Test: Really difficult to test this scenario without proper test
setup. Since we have no tests for station.h and are about to replace
the entire file soon, I suggest we move forward and write a
corresponding regression test once the new client landed.
Filed NET-898

NET-888 #done

Change-Id: I4d880f6411a04606deffadc13ab020a2586b6885