[dash] fix validation of nodes length
Previously the condition asserted that the buffer to decode was exactly
the declared length. This was more aggressive than needed, especially as
there are valid reasons for the buffer being larger, such as padding.
The new condition asserts that the buffer is at least as large as its
declared length.

ZX-1920 #fix

Change-Id: I7cbd0c5a6782dce4872bcf66c03b841124999d33