[images] Fix amber_index contents
The amber index gets the file names of meta.far files in the build,
not the merkleroots (as in the pkgsvr index) nor the names of files
containing merkleroots.  We'll probably change amber later so that
it consumes the pkgsvr index directly (i.e. finds each meta.far as
a blob via its merkleroot).

BLD-168 #done

Change-Id: I3262e46b8725bab9d9be31e8c53688610f63dfec