[util][utf] Start a UTF conversion library.
Start a basic UTF conversion library.  Right now, it only converts from UTF-16
to UTF-8, but it can be extended as needs arise.  For now, I plan to use it to
convert strings in USB descriptors (which are spec'ed to be UTF16) to UTF8
(which seems to be the current defacto standard for the rest of Fuchsia)

Change-Id: Ie2bcccb82ec00680752a2fac4d937332bd52fc95