[fit] Define fit::function.
Introduce a new library, libfit, which is intended to contain a
small number of abstractions that serve as a kind of "annex" to
the C++ 14 standard library.

This code has been derived from fbl::Function with the following
changes:
- adopt an API style similar to that of std::function
- use hand-rolled vtables to improve code generation
- add missing target() member function
- drop AllocChecker overloads

Added an empty file to make the linker happy, as was previously
done for libfbl way back when it too was header-only.

Change-Id: I6412a1827e34a9ac63444a9bc07c692ec52d763e